+++ src/instance/methods/behavior.js
@@ -11,5 +11,9 @@ export async function addBehavior(name, options = {}) {
  * Attach a behavior to the farmOS-map instance.
  */
 export async function attachBehavior(behavior, options = {}) {
+  // Wait until all the default behaviors are done being attached
+  await this.defaultBehaviorsAttached.finally();
+
+  // Attach the requested behavior
   return behavior.attach(this, options);
 }